+++ native/user_c_modules/openbricks/micropython.cmake
@@ -16,6 +16,8 @@ target_sources(usermod_openbricks INTERFACE
     ${CMAKE_CURRENT_LIST_DIR}/st_servo_core.c
     ${CMAKE_CURRENT_LIST_DIR}/st_move_core.c
     ${CMAKE_CURRENT_LIST_DIR}/imu_yaw_core.c
+    ${CMAKE_CURRENT_LIST_DIR}/icm45686.c
+    ${CMAKE_CURRENT_LIST_DIR}/icm45686_core.c
     ${CMAKE_CURRENT_LIST_DIR}/st_button_core.c
     ${CMAKE_CURRENT_LIST_DIR}/trajectory.c
     ${CMAKE_CURRENT_LIST_DIR}/trajectory_core.c